- client now listens response from server (added accept cycle, rewrited connect) - client&server: accept timeout can be set - client sends vector of byte (instead vector of int) - updated README, added comments to code - small fixes

+++ main_client.cpp
@@ -6,8 +6,8 @@
 
 using namespace std;
 
-vector<int> getVector(const std::string& str) {
-    vector<int> v;
+vector<uint8_t> getVector(const std::string& str) {
+    vector<uint8_t> v;
     size_t pos = str.find(',');
     size_t start = 0;
     while (start != string::npos) {
